requirements.txt
click >= 8.1
GitPython >= 3.1